MIDI encoder fix

+++ lib/SongConverter.cpp
@@ -341,8 +341,8 @@ class MIDIEncoder
 
     void _sendContinuedValue(uint32_t val)
     {
-        uint32_t send[3] = {};
-        uint32_t* ptr = nullptr;
+        uint8_t send[3] = {};
+        uint8_t* ptr = nullptr;
         if (val >= 0x4000)
         {
             ptr = &send[0];
